Fragmented the stress package into multiple files
